feat(attachments): backfill op + drop is_inline (PRD #502, Brief 2)
Closes PRD #502 in full. Replaces the Gmail-API-dependent
recompute-attachment-inline op with a filename-only
recompute-attachment-decoration op, drops the now-dead is_inline
column, and removes the isInline field from contracts/schema/mappers/
repos/capture/tests.

- New ops:recompute-attachment-decoration command (dry-run default,
  --execute, summary JSON, no Gmail API — works from already-stored
  filename + mime_type)
- Deletes apps/worker/src/ops/recompute-attachment-inline.ts and its
  test (500 + 592 LOC)
- Migration 0068 drops message_attachments.is_inline
- Migrator sentinel flips from is_inline to id (stable PK)
- isInline removed from MessageAttachmentRecord, Drizzle schema, row
  mapper, domain repository, gmail-capture write, and all tests

Historical Karen-shaped rows heal as soon as the architect runs
ops:recompute-attachment-decoration --execute against prod.

PRD: #502
Builds on: #503
