+++ apps/gmail-capture/test/internal-attachments-route.test.ts
@@ -91,7 +91,6 @@ async function seedAttachment(input: {
         sizeBytes: input.sizeBytes,
         storageKey: input.storageKey,
         isDecoration: false,
-        isInline: false,
       },
     ],
   );
