+++ packages/db/src/mappers.ts
@@ -755,7 +755,6 @@ export function mapMessageAttachmentRow(
     sizeBytes: row.sizeBytes,
     storageKey: row.storageKey,
     isDecoration: row.isDecoration,
-    isInline: row.isInline,
     createdAt: row.createdAt.toISOString(),
   });
 }
@@ -775,7 +774,6 @@ export function mapMessageAttachmentToInsert(
     sizeBytes: parsed.sizeBytes,
     storageKey: parsed.storageKey,
     isDecoration: parsed.isDecoration,
-    isInline: parsed.isInline,
     createdAt: toDate(parsed.createdAt),
   };
 }
